# remove unused HGVertex member variables