Remove sets populated during vertex construction but never used.
* Saves > 82 MB RAM in C++ version.
* A decent speedup `Creating unique names and vertices`:
  * \>23% C++ on lab2 (4.3s -> 3.48s)
  * \~44% Python on noreaster (17.6s -> 12.2s)