@yakra 2021-01-11
I assume so. I followed the blame back to when they were first introduced, but didn't take too close a look at the surrounding code to see if they were used.