@yakra 2021-01-11
I've got some upcoming C++ changes (\~25% speedup writing subgraphs) that follow from this commit.
If this is something really worth saving, I could cherrypick or do an interactive rebase.
Or, if the need should ever arise in the future, it should be easy enough to add back in.